Consider application/zip as jar file
The libmagic can detect jar files as zip file. Under the hood, this files
as indeed a kind of zip file. But this can cause issue for us when
detecting if the given file by the user is a jar or not. For this
reason, the is_jar function now considers the "application/zip" mime type as
jar files.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>