Text extractor interface.
The text extractor interface goal is to abstract for the user all the
logic necessary to extract text from a given gazette. The interface has
methods to extract the text and metadata of the given gazette them back
to the disk in files. It also has the methods to load these extract data
into memory to allow the user manipulation.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>